new file:   config/message_center.sql 	modified:   config/parameters.php 	modified:   config/paths.php 	modified:   controllers/error.php 	new file:   controllers/mail.php 	modified:   index.php 	modified:   libs/bootstrap.php 	modified:   libs/check.php 	modified:   libs/controller.php 	modified:   libs/database.php 	new file:   libs/functions.php 	new file:   libs/functions/functions.php 	modified:   libs/views.php 	new file:   models/error.php 	modified:   models/message.php 	modified:   models/user.php 	new file:   public/css/custom.css 	new file:   public/img/logo.jpg 	new file:   public/js/custom.js 	new file:   test.php 	new file:   views/default/error/error.php 	new file:   views/default/footer.php 	new file:   views/default/header.php 	new file:   views/default/login/index.php 	new file:   views/default/mail/components/message.php 	new file:   views/default/mail/components/table-cells.php 	new file:   views/default/mail/components/table.php 	new file:   views/default/mail/compose.php 	new file:   views/default/mail/inbox.php 	new file:   views/default/mail/sentbox.php 	new file:   views/default/mail/trash.php 	new file:   views/default/main/index.php 	new file:   views/default/sidebar.php

+++ libs/database.php
@@ -17,51 +17,55 @@ class Database extends PDO {
         $this -> table = $table;
     }
 
-   
     /*
      * setCondition($cond)
-     * 
-     * * be careful and use this function carefully 
-     * * as it could be used in SQL injection 
+     *
+     * * be careful and use this function carefully
+     * * as it could be used in SQL injection
      * * NEVER pass arguments directly from the user ($_GET OR $_POST OR $_COOKIES)
-     * 
+     *
      * */
     public function setCondition($cond) {
         $this -> condition = $cond;
     }
 
-    public function getResult($command = "SELECT") {
-        if ($command == "SELECT") {
-            $sql = "SELECT * FROM " . $this->table . " WHERE " . $this->condition . " ;";
+    public function getResult($query = NULL, $command = "SELECT") {
+        if ($query === NULL) {
+            if ($command == "SELECT") {
+                $sql = "SELECT * FROM " . $this -> table . " WHERE " . $this -> condition . " ;";
+                $sSQL = 'SET CHARACTER SET utf8';
+                $this -> exec($sSQL);
+                $stm = $this -> query($sql);
+                // $stm -> execute(array(":condition" => $this -> condition, ":table" => $this -> table));
+                // print_r ($stm -> fetchAll());
+                // echo $this->condition . $this->table;
+                return $stm -> fetchAll();
+            } else if ($command == "DELETE") {
+                $sql = "DELETE FROM " . $this -> table . " WHERE " . $this -> condition . " ;";
+                $sSQL = 'SET CHARACTER SET utf8';
+                $this -> exec($sSQL);
+                $stm = $this -> query($sql);
+                $result = $stm -> rowCount() . "rows deleted.";
+                return $result;
+
+            }
+        } else {
             $sSQL = 'SET CHARACTER SET utf8';
             $this -> exec($sSQL);
-            $stm = $this -> query($sql);
-            // $stm -> execute(array(":condition" => $this -> condition, ":table" => $this -> table));
-            // print_r ($stm -> fetchAll());
-            // echo $this->condition . $this->table;
+            $stm = $this -> query($query);
             return $stm -> fetchAll();
-        } else if ($command == "DELETE") {
-            $sql = "DELETE FROM " . $this->table . " WHERE " . $this->condition . " ;";
-            $sSQL = 'SET CHARACTER SET utf8';
-            $this -> exec($sSQL);
-            $stm = $this -> query($sql);
-            $result = $stm -> rowCount()."rows deleted.";
-            return $result;
-            
         }
     }
-    
-    public function getId($id , $table = Null )
-    {
-        if(isset($table) && !empty($table))
-        {
-            $this ->setCondition(" id = " . $id );
-            
-            $result = $this->getResult("SELECT");
+
+    public function getId($id, $table = Null) {
+        if (isset($table) && !empty($table)) {
+            $this -> setCondition(" id = " . $id);
+
+            $result = $this -> getResult("SELECT");
             return $result;
         } else {
-            $curTable = $this->table;
-            $this->table = $table ;
+            $curTable = $this -> table;
+            $this -> table = $table;
             $result = $this -> getResult("SELECT");
         }
     }
